backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
